Provide product_info non-existent feature bypas
The code of the pages/product-info javascript was updated to include some additional features; however, for systems that do not have the associated/sought code, javascript errors occur.  The code added to this commit is intended to prevent the javascript error, but also to allow the new code to continue to function.